+++ CHANGELOG.md
@@ -1,3 +1,30 @@
+## [0.50.0](https://github.com/propeller-heads/tycho-simulation/compare/0.49.1...0.50.0) (2024-11-29)
+
+
+### Features
+
+* move balances to init - it's now mandatory ([f9b9131](https://github.com/propeller-heads/tycho-simulation/commit/f9b9131c1b2e10a68e50ae62ad17322268425506))
+
+
+### Bug Fixes
+
+* Don't use unwrap_or_default if balance not found ([a6c0a37](https://github.com/propeller-heads/tycho-simulation/commit/a6c0a3791bdef52c8a3fcf177714175150552ff6))
+
+## [0.49.1](https://github.com/propeller-heads/tycho-simulation/compare/0.49.0...0.49.1) (2024-11-29)
+
+
+### Bug Fixes
+
+* Use u256_to_f64 when calculating prices ([49a8a31](https://github.com/propeller-heads/tycho-simulation/commit/49a8a3121b8d9a7bd57bf6732b4b2c1341a720e2))
+
+## [0.49.0](https://github.com/propeller-heads/tycho-simulation/compare/0.48.0...0.49.0) (2024-11-29)
+
+
+### Features
+
+* Change type of adapter_contract_path to PathBuf ([d78a761](https://github.com/propeller-heads/tycho-simulation/commit/d78a761353a1827d0a259657acd9980cffed70be))
+* The adapter file path needs to be relative to this crate ([8fefefd](https://github.com/propeller-heads/tycho-simulation/commit/8fefefdcb35857b861b4b52e52e9791634ca3bbb))
+
 ## [0.48.0](https://github.com/propeller-heads/tycho-simulation/compare/0.47.3...0.48.0) (2024-11-28)